Partially update to Ryu
