+++ test/runtests.jl
@@ -1,15 +1,12 @@
 using Showoff
 using Test
 using Dates
-if isdefined(Base, :Grisu)
-    const Grisu = Base.Grisu
-else
-    import Grisu
-end
 
 @testset "Internals" begin
     @test Showoff.@grisu_ccall(1, 2, 3) === nothing
-    @test Showoff.grisu(1.0, Grisu.SHORTEST, 2) == (1, 1, false, Grisu.DIGITS)
+    if isdefined(Showoff, :Grisu)
+        @test Showoff.grisu(1.0, Showoff.Grisu.SHORTEST, 2) == (1, 1, false, Showoff.Grisu.DIGITS)
+    end
 
     let x = [1.0, Inf, 2.0, NaN]
         @test Showoff.concrete_minimum(x) == 1.0
